Fixed some backend things

+++ src/main/java/jl/corona/controller/ContinentController.java
@@ -35,7 +35,7 @@ public class ContinentController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteCountry(@PathVariable long id) {
+    public void deleteCountry(@PathVariable int id) {
         continentRepository.deleteContinent(id);
     }
 }
